Proposal for updating the overview page of the developer perspective
This is the inital proposal for updating the overview page of the developer console. The goals of this proposal include providing the user with a high level description of the main use cases of the dev console, the main types of users who use it, latest release information, and provide new designers with a better onboarding experience to the dev console.

Another goal of this overview page is to provide the user with links to the latest design work but not have it be the main focus of the overivew page.

Ideally if the user wanted to look at older design work, they could access those pages from the release section. We can provide the user with a link to that section from this page.