Cleaning up the Developer page so it's just releases and a couple other edits
Cleaned up the developer page so it features just the release info, similar to the administrtor page. I also updated the nav order so the nav items are now in alphabetical order. I also added an overview section to the nav items so you can find overview, persona, and use case information for each perspective.